ref(core): consolidate inputGroup component files (#105571)
Moves all content from inputGroup.chonk.tsx into inputGroup.tsx. Removes
"Chonk" and "chonk" prefixes from all styled components and functions
(ChonkStyledInput, ChonkStyledTextArea, ChonkStyledLeadingItemsWrap,
ChonkStyledTrailingItemsWrap, chonkInputStyles, chonkItemsPadding).
