+++ static/app/components/core/input/inputGroup.tsx
@@ -6,18 +6,86 @@ import {
   useRef,
   useState,
 } from 'react';
+import {css} from '@emotion/react';
 import styled from '@emotion/styled';
 
 import type {InputProps} from 'sentry/components/core/input';
-import {Input as _Input} from 'sentry/components/core/input';
-import {
-  ChonkStyledInput,
-  ChonkStyledLeadingItemsWrap,
-  ChonkStyledTextArea,
-  ChonkStyledTrailingItemsWrap,
-} from 'sentry/components/core/input/inputGroup.chonk';
+import {Input as _Input, Input as CoreInput} from 'sentry/components/core/input';
 import type {TextAreaProps} from 'sentry/components/core/textarea';
-import {TextArea as _TextArea} from 'sentry/components/core/textarea';
+import {
+  TextArea as _TextArea,
+  TextArea as CoreTextArea,
+} from 'sentry/components/core/textarea';
+import {space} from 'sentry/styles/space';
+import type {FormSize, StrictCSSObject, Theme} from 'sentry/utils/theme';
+
+interface InputStyleProps {
+  leadingWidth?: number;
+  size?: FormSize;
+  trailingWidth?: number;
+}
+
+const InputItemsWrap = styled('div')`
+  display: grid;
+  grid-auto-flow: column;
+  align-items: center;
+  gap: ${space(1)};
+
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+`;
+
+const itemsPadding = {
+  md: 8,
+  sm: 6,
+  xs: 4,
+} satisfies Record<NonNullable<InputStyleProps['size']>, number>;
+
+const inputStyles = ({
+  leadingWidth,
+  trailingWidth,
+  size = 'md',
+  theme,
+}: InputStyleProps & {theme: Theme}): StrictCSSObject => css`
+  ${leadingWidth &&
+  css`
+    padding-left: calc(
+      ${theme.form[size].paddingLeft}px + ${itemsPadding[size]}px + ${leadingWidth}px
+    );
+  `}
+
+  ${trailingWidth &&
+  css`
+    padding-right: calc(
+      ${theme.form[size].paddingRight}px + ${itemsPadding[size]}px + ${trailingWidth}px
+    );
+  `}
+`;
+
+const StyledInput = styled(CoreInput)<InputStyleProps>`
+  ${inputStyles}
+`;
+
+const StyledTextArea = styled(CoreTextArea)<InputStyleProps>`
+  ${inputStyles}
+`;
+
+const StyledLeadingItemsWrap = styled(InputItemsWrap)<{
+  size: NonNullable<InputStyleProps['size']>;
+  disablePointerEvents?: boolean;
+}>`
+  left: ${p => p.theme.form[p.size].paddingLeft + 1}px;
+  ${p => p.disablePointerEvents && `pointer-events: none;`}
+`;
+
+const StyledTrailingItemsWrap = styled(InputItemsWrap)<{
+  size: NonNullable<InputStyleProps['size']>;
+  disablePointerEvents?: boolean;
+}>`
+  right: ${p => p.theme.form[p.size].paddingRight + 1}px;
+  ${p => p.disablePointerEvents && `pointer-events: none;`}
+`;
 
 interface InputContext {
   /**
@@ -145,7 +213,7 @@ function LeadingItems({children, disablePointerEvents, ...props}: InputItemsProp
   }, [children, setLeadingWidth, size]);
 
   return (
-    <InputLeadingItemsWrap
+    <StyledLeadingItemsWrap
       ref={ref}
       size={size}
       disablePointerEvents={disabled || disablePointerEvents}
@@ -153,7 +221,7 @@ function LeadingItems({children, disablePointerEvents, ...props}: InputItemsProp
       {...props}
     >
       {children}
-    </InputLeadingItemsWrap>
+    </StyledLeadingItemsWrap>
   );
 }
 
@@ -180,7 +248,7 @@ function TrailingItems({children, disablePointerEvents, ...props}: InputItemsPro
   }, [children, setTrailingWidth, size]);
 
   return (
-    <InputTrailingItemsWrap
+    <StyledTrailingItemsWrap
       ref={ref}
       size={size}
       disablePointerEvents={disabled || disablePointerEvents}
@@ -188,7 +256,7 @@ function TrailingItems({children, disablePointerEvents, ...props}: InputItemsPro
       {...props}
     >
       {children}
-    </InputTrailingItemsWrap>
+    </StyledTrailingItemsWrap>
   );
 }
 
@@ -203,11 +271,3 @@ const InputGroupWrap = styled('div')<{disabled?: boolean}>`
   position: relative;
   ${p => p.disabled && `color: ${p.theme.disabled};`};
 `;
-
-const StyledInput = ChonkStyledInput;
-
-const StyledTextArea = ChonkStyledTextArea;
-
-const InputLeadingItemsWrap = ChonkStyledLeadingItemsWrap;
-
-const InputTrailingItemsWrap = ChonkStyledTrailingItemsWrap;
